// src/Types/auth.ts
export type SignupFormValues = {
  email: string;
  password: string;
  fullName: string;
  profilePic?: string|null;
  profilePicUrl?:string|null;
};

export type LoginFormValues = {
  email: string;
  password: string
};

export type UserProfile = {
  id: string;
  email:string;
  full_name: string;
  profile_pic: string;
  role: "student" | "teacher" | "admin";
  bio: string | null;
};